FIX: Resolved breaking change causing 502 bad gateway
-The pom file had incorrectly included the maven-assembly-plugin into the pluginManagement section, causing the assembly jobs to not run
