+++ pom.xml
@@ -137,8 +137,8 @@
             <scope>test</scope>
         </dependency>
         <!-- END test dependencies -->
-
     </dependencies>
+
     <dependencyManagement>
         <dependencies>
             <dependency>
@@ -152,9 +152,9 @@
     </dependencyManagement>
 
     <build>
-        <pluginManagement>
             <plugins>
                 <plugin>
+                    <!-- IMPORTANT NOTE https://stackoverflow.com/questions/54279662/maven-jar-not-including-main-in-manifest -->
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-assembly-plugin</artifactId>
                     <executions>
@@ -179,15 +179,54 @@
                     </executions>
                 </plugin>
 
+                <!-- Build an executable JAR -->
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-jar-plugin</artifactId>
+                    <version>3.2.0</version>
+                    <configuration>
+                        <archive>
+                            <manifest>
+                                <mainClass>edu.bu.cs673.stockportfolio.StockportfolioApplication</mainClass>
+                            </manifest>
+                        </archive>
+                    </configuration>
+                </plugin>
+
+                <!-- START unit test dependency -->
                 <plugin>
                     <artifactId>maven-surefire-plugin</artifactId>
                     <version>3.0.0-M5</version>
                     <!-- This configuration is required to avoid timeout errors in Jenkins -->
                     <configuration>
-                        <forkedProcessExitTimeoutInSeconds>300</forkedProcessExitTimeoutInSeconds>
-                        <forkCount>0</forkCount>
+                        <excludes>
+                            <include>**/*IT.java</include>
+                        </excludes>
                     </configuration>
                 </plugin>
+                <!-- END unit test dependency -->
+
+                <!-- START integration test dependency run with mvn failsafe:integration-test -->
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-failsafe-plugin</artifactId>
+                    <configuration>
+                        <forkedProcessExitTimeoutInSeconds>300</forkedProcessExitTimeoutInSeconds>
+                        <forkCount>0</forkCount> <!-- TRY TO REMOVE -->
+                        <includes>
+                            <include>**/*IT.java</include>
+                        </includes>
+                    </configuration>
+                    <executions>
+                        <execution>
+                            <goals>
+                                <goal>integration-test</goal>
+                                <goal>verify</goal>
+                            </goals>
+                        </execution>
+                    </executions>
+                </plugin>
+                <!-- END integration test dependency -->
 
                 <!-- START spotbugs plugin https://spotbugs.readthedocs.io/en/latest/maven.html -->
                 <plugin>
@@ -211,7 +250,6 @@
                     <version>2.5.0</version>
                 </plugin>
             </plugins>
-        </pluginManagement>
     </build>
 
 </project>
