--- src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: edu.bu.cs673.stockportfolio.StockportfolioApplication
-
