+++ src/test/java/edu/bu/cs673/stockportfolio/integrationtests/StockPortfolioIT.java
@@ -1,4 +1,4 @@
-package edu.bu.cs673.stockportfolio.unittests;
+package edu.bu.cs673.stockportfolio.integrationtests;
 
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.platform.suite.api.SelectPackages;
@@ -6,8 +6,8 @@ import org.junit.runner.RunWith;
 
 @RunWith(JUnitPlatform.class)
 @SelectPackages({
-        "edu.bu.cs673.stockportfolio.unittests"
+        "edu.bu.cs673.stockportfolio.integrationtests"
 })
-public class UnitTests {
+public class StockPortfolioIT {
     // This class remains empty, it is used only as a holder for the above annotations
 }
