+++ src/test/java/edu/bu/cs673/stockportfolio/integrationtests/dashboards/StockDashboardIT.java
@@ -1,11 +1,12 @@
-package edu.bu.cs673.stockportfolio.integrationtests.homepage;
+package edu.bu.cs673.stockportfolio.integrationtests.dashboards;
 
 import edu.bu.cs673.stockportfolio.integrationtests.homepage.HomePage;
 import edu.bu.cs673.stockportfolio.integrationtests.login.LoginPage;
-import edu.bu.cs673.stockportfolio.integrationtests.utilityPages.ResultPage;
 import edu.bu.cs673.stockportfolio.integrationtests.signup.SignupPage;
+import edu.bu.cs673.stockportfolio.integrationtests.utilityPages.ResultPage;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.jupiter.api.*;
+import org.openqa.selenium.Dimension;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
@@ -15,30 +16,29 @@ import org.springframework.boot.web.server.LocalServerPort;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**********************************************************************************************************************
- * Test user story: "As a user, I can signup, login, and upload my portfolio."
+ * Test user story: "As a user, I can signup, login, and view the stock dashboard for my portfolio."
  *********************************************************************************************************************/
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class UploadPortfolioTest {
+public class StockDashboardIT {
 
     @LocalServerPort
     private int port;
-
     private static WebDriver driver;
-    private String baseURL;
+    private String baseUrl;
     private SignupPage signupPage;
     private LoginPage loginPage;
     private HomePage homePage;
     private ResultPage resultPage;
 
     @BeforeAll
-    static void beforeAll() {
+    public static void beforeAll() {
         WebDriverManager.chromedriver().setup();
     }
 
     @BeforeEach
     public void beforeEach() {
         driver = new ChromeDriver();
-        baseURL = "http://localhost:" + port;
+        baseUrl = "http://localhost:" + port;
         signupPage = new SignupPage(driver);
         loginPage = new LoginPage(driver);
         homePage = new HomePage(driver);
@@ -46,40 +46,33 @@ public class UploadPortfolioTest {
     }
 
     @Test
-    @DisplayName("Test portfolio upload.")
-    public void testPortfolioUpload() {
+    @DisplayName("Test stock dashboard link")
+    public void testMarketCapLink() {
         String filePath = "/Users/mlewis/Downloads/TestPortfolio.csv";
-        String testTickerOne = "GS";
-        String testTickerTwo = "FB";
-        String testCompanyNameOne = "Goldman Sachs Group, Inc.";
-        String testCompanyNameTwo = "Facebook Inc - Class A";
 
-        driver.get(baseURL + "/signup");
-        signupPage.signup("money@spd.com", "John", "10DigitPassword!");
+        driver.get(baseUrl + "/signup");
+        signupPage.signup("stocks@spd.com", "buyside", "buyside1234");
 
-        driver.get(baseURL + "/login");
-        loginPage.login(driver, "John", "10DigitPassword!");
+        driver.get(baseUrl + "/login");
+        loginPage.login(driver, "wall", "wallStreet");
 
-        // Go to homepage and upload file
-        driver.get(baseURL + "/home");
+        // First upload a portfolio to ensure full integration testing with IEX Cloud
+        driver.get(baseUrl + "/home");
         homePage.clickUploadPortfolio(driver, filePath);
         homePage.clickUploadPortfolioButton(driver);
 
         boolean result = resultPage.isSuccessMessageDisplayed(driver);
         resultPage.clickNavLink(driver);
 
-        driver.get(baseURL + "/home");
-        String actualTickerOne = homePage.find(driver, testTickerOne);
-        String actualTickerTwo = homePage.find(driver, testTickerTwo);
-        String actualCompanyNameOne = homePage.find(driver, testCompanyNameOne);
-        String actualCompanyNameTwo = homePage.find(driver, testCompanyNameTwo);
+        driver.get(baseUrl + "/home");
+        homePage.clickSectorBreakdown(driver);
+        String currentUrl = driver.getCurrentUrl();
 
-        assertAll("Upload portfolio",
+        assertAll("Stock Breakdown Dashboard",
                 () -> assertTrue(result, "Portfolio upload failed."),
-                () -> assertEquals(testTickerOne, actualTickerOne, "GS symbol is incorrect."),
-                () -> assertEquals(testTickerTwo, actualTickerTwo, "FB symbol is incorrect."),
-                () -> assertEquals(testCompanyNameOne, actualCompanyNameOne, "Company name is incorrect."),
-                () -> assertEquals(testCompanyNameTwo, actualCompanyNameTwo, "Company name is incorrect."));
+                () -> assertEquals(baseUrl + "/sector_breakdown", currentUrl,
+                        "Incorrect endpoint for stock dashboard")
+        );
     }
 
     @AfterEach
@@ -87,6 +80,7 @@ public class UploadPortfolioTest {
         if (driver != null) {
             driver.quit();
         }
+
         driver = null;
     }
 }
