+++ src/test/java/edu/bu/cs673/stockportfolio/integrationtests/security/AuthorizationIT.java
@@ -1,7 +1,5 @@
-package edu.bu.cs673.stockportfolio.integrationtests.signup;
+package edu.bu.cs673.stockportfolio.integrationtests.security;
 
-import edu.bu.cs673.stockportfolio.integrationtests.homepage.HomePage;
-import edu.bu.cs673.stockportfolio.integrationtests.login.LoginPage;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.jupiter.api.*;
 import org.openqa.selenium.WebDriver;
@@ -10,20 +8,19 @@ import org.openqa.selenium.chrome.ChromeOptions;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 /**********************************************************************************************************************
- * Test user story: "As a user, I can signup, login, and logout."
+ * Test user story: "As a security engineer, I want to protect the product from unauthorized access."
  *********************************************************************************************************************/
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class StandUpAndTearDownTest {
+public class AuthorizationIT {
 
     @LocalServerPort
     private int port;
 
     private static WebDriver driver;
     private String baseURL;
-    private SignupPage signupPage;
-    private LoginPage loginPage;
-    private HomePage homePage;
 
     @BeforeAll
     static void beforeAll() {
@@ -34,27 +31,19 @@ public class StandUpAndTearDownTest {
     public void beforeEach() {
         driver = new ChromeDriver();
         baseURL = "http://localhost:" + port;
-        signupPage = new SignupPage(driver);
-        loginPage = new LoginPage(driver);
-        homePage = new HomePage(driver);
     }
 
     @Test
-    @DisplayName("Test signup, login, and logout.")
-    public void testSignUpLoginAndLogout() {
-
+    @DisplayName("Test restricted access for unauthorized user.")
+    public void unauthorizedAccess() {
         driver.get(baseURL + "/signup");
-        signupPage.signup("money@spd.com", "John", "10DigitPassword!");
+        assertEquals("Sign Up", driver.getTitle());
 
         driver.get(baseURL + "/login");
-        loginPage.login(driver, "John", "10DigitPassword!");
+        assertEquals("Login", driver.getTitle());
 
         driver.get(baseURL + "/home");
-        homePage.logout(driver);
-
-        driver.get(baseURL + "/home");
-        String currentUrl = driver.getCurrentUrl();
-        Assertions.assertNotEquals(baseURL + "/home", currentUrl);
+        assertEquals("Login", driver.getTitle());
     }
 
     @AfterEach
