+++ src/test/java/edu/bu/cs673/stockportfolio/integrationtests/signup/StandUpAndTearDownIT.java
@@ -1,5 +1,7 @@
-package edu.bu.cs673.stockportfolio.integrationtests.security;
+package edu.bu.cs673.stockportfolio.integrationtests.signup;
 
+import edu.bu.cs673.stockportfolio.integrationtests.homepage.HomePage;
+import edu.bu.cs673.stockportfolio.integrationtests.login.LoginPage;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.jupiter.api.*;
 import org.openqa.selenium.WebDriver;
@@ -8,19 +10,20 @@ import org.openqa.selenium.chrome.ChromeOptions;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
 /**********************************************************************************************************************
- * Test user story: "As a security engineer, I want to protect the product from unauthorized access."
+ * Test user story: "As a user, I can signup, login, and logout."
  *********************************************************************************************************************/
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class AuthorizationTest {
+public class StandUpAndTearDownIT {
 
     @LocalServerPort
     private int port;
 
     private static WebDriver driver;
     private String baseURL;
+    private SignupPage signupPage;
+    private LoginPage loginPage;
+    private HomePage homePage;
 
     @BeforeAll
     static void beforeAll() {
@@ -31,19 +34,27 @@ public class AuthorizationTest {
     public void beforeEach() {
         driver = new ChromeDriver();
         baseURL = "http://localhost:" + port;
+        signupPage = new SignupPage(driver);
+        loginPage = new LoginPage(driver);
+        homePage = new HomePage(driver);
     }
 
     @Test
-    @DisplayName("Test restricted access for unauthorized user.")
-    public void unauthorizedAccess() {
+    @DisplayName("Test signup, login, and logout.")
+    public void testSignUpLoginAndLogout() {
+
         driver.get(baseURL + "/signup");
-        assertEquals("Sign Up", driver.getTitle());
+        signupPage.signup("money@spd.com", "John", "10DigitPassword!");
 
         driver.get(baseURL + "/login");
-        assertEquals("Login", driver.getTitle());
+        loginPage.login(driver, "John", "10DigitPassword!");
 
         driver.get(baseURL + "/home");
-        assertEquals("Login", driver.getTitle());
+        homePage.logout(driver);
+
+        driver.get(baseURL + "/home");
+        String currentUrl = driver.getCurrentUrl();
+        Assertions.assertNotEquals(baseURL + "/home", currentUrl);
     }
 
     @AfterEach
